decision to use the deploy user for more stuff. explanation below.
so far the actual deployment by travis when done as root and later all files where chowned to www-data. this sucks in multiple ways: we should not have to use root for that and there's no reason for using www-data as uwsgi is doing the acutal work (besides nginx running as www-data serving static files).
all this is changed in the following commits and has quite some implications.
this first commit moves the project directory from /var/www to $HOME of the deploy user