new separate playbook for getting TLS certificates.
why separate? we need TLS certificates for the nginx. however, to get these certificates, we need nginx. it's the chicken or egg dilemma. we do not want to turn TLS off, get certificates (or keep current ones) and turn TLS on again every time we deploy the configuration. an alternative would have been using conditionals to check if the certificates already exist. if yes, we could have skipped the TLS stuff. however, conditionals are not that convenient using ansible.
so instead this playbook is meant to be run first and probably only once. after that the certificates exist and the normal configuration playbook can be applied as often as you want without turning TLS off temporarily.