include nginx wordpress config.
bad situation: this repo contains only the main strassengezwitscher stuff. the wordpress blog is probably temporary and not part of this repo. however, as it is running under one of the main domains, we need to include at least the reference to its nginx config. not really happy about it.