moved location where nginx expects static files
corresponding settings will be commited in strassengezwitscher repo. as the static files are no longer inside the django directory, www-data does not need to be able to read it. therefore the access to that directory could be limited (previous commit). this leads to a securer storage of sensitive settings within the django directory.